Fixed integer underflow bug when jumping during scrolling
I didn't take care of situation when user scrolled to
bottom of file and then jumped immedaitely to the beginning.

This resulted in optimal starting line being determined to be
at -10, which resulted in integer underflow before this fix.

The integer underflow resulted in the view jumping to
end of file before returning back to top when user
moved cursor.