Commit `Cargo.lock` file
The `Cargo.lock` file is recommended to be checked in for executables.
It allows building and installing the application through cargo in a way
that references a common set of dependencies.

This is also important for distributions like Fedora, or NixOs, that use
the `cargo vendor` command to package the actual version that is
released rather than everyone having a package that can essentialy
behave differently in non obvious ways.

Reference:
- https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html